services/reco-builder/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
numpy==1.26.2
sentence-transformers==2.2.2
huggingface-hub==0.19.4
faiss-cpu==1.7.4
python-dotenv==1.0.0
requests==2.31.0
torch==2.0.1
transformers==4.35.2
psycopg2-binary==2.9.9
minio==7.2.0
